table_producer upgraded
table producer can now pull all the comments above functions in any
format and also grab the associated function. I still need to save this
data but now that I've got it ready I'm in great shape to get table
producer fully functional next time I sit down to code this.

After code producer I will need to make code updater and maybe a cursory
search program to demo, which could be valuable from a show-off
perspective.  I know that this method of approaching search is not going
to yield a system optimal from a performance perspective, but I don't
have the time or the team to do this the right way.

Jury rig it and wing it from there right????